requirements.txt
python-dotenv==1.0.1
deepl==1.21.0
python-telegram-bot==21.10
requests==2.32.3
elevenlabs==1.50.7

